feat: Refactor recommendation engine with a 50/20/30 distribution strategy, introduce user-hidden images, and implement a scheduled preference decay command.

+++ app/Jobs/UpdateUserPreferencesJob.php
@@ -21,7 +21,7 @@ class UpdateUserPreferencesJob implements ShouldQueue
     /**
      * Create a new job instance.
      */
-    public function __construct(string $userId, string $imageId, int $weightAdjustment = 1)
+    public function __construct(string $userId, string $imageId, float $weightAdjustment = 1.0)
     {
         $this->userId = $userId;
         $this->imageId = $imageId;
